[oneDNN-upstream-merge pipeline]

+++ src/gpu/intel/jit/ir/linear_expr.hpp
@@ -42,15 +42,15 @@ expr_t to_linear(const expr_t &_e);
 
 // Returns the max power of two divisor of an expression. The expression must
 // be convertable to linear_t.
-int linear_max_pow2_divisor(const expr_t &e);
+int64_t linear_max_pow2_divisor(const expr_t &e);
 
 // Divides an expression by a constant. The expression must be convertable to
 // linear_t.
-expr_t linear_div(const expr_t &e, int factor);
+expr_t linear_div(const expr_t &e, int64_t factor);
 
 // Simplifies a modulus of an expression by a constant. The expression must be
 // convertable to linear_t.
-expr_t simplify_linear_mod(const expr_t &e, int factor);
+expr_t simplify_linear_mod(const expr_t &e, int64_t factor);
 
 // Returns the base and the increments of linear expression `expr` when
 // incrementing `idxs[i]` by 1:
